+++ modules/foundation/src/main/kotlin/com/teamwizardry/librarianlib/foundation/util/FoundationSerializerExceptions.kt
@@ -1,13 +1,13 @@
 package com.teamwizardry.librarianlib.foundation.util
 
-open class FoundationSerializationException: RuntimeException {
+open class FoundationSerializerException: RuntimeException {
     constructor(): super()
     constructor(message: String): super(message)
     constructor(message: String, cause: Throwable?): super(message, cause)
     constructor(cause: Throwable): super(cause)
 }
 
-class InvalidSerializedClassException: FoundationSerializationException {
+class InvalidSerializedClassException: FoundationSerializerException {
     constructor(): super()
     constructor(message: String): super(message)
     constructor(message: String, cause: Throwable?): super(message, cause)
